% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta.R
\name{fbetalik}
\alias{fbetalik}
\title{Beta Likelihood Function}
\usage{
fbetalik(data, a, b, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{a}{shape parameter alpha}

\item{b}{shape parameter beta}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Beta Likelihood Function
}

