% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{fflik}
\alias{fflik}
\title{F Likelihood Function}
\usage{
fflik(data, df1, df2, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{df1}{degrees of freedom}

\item{df2}{degrees of freedom}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
F Likelihood Function
}
