% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta.R
\name{likbeta}
\alias{likbeta}
\title{Beta Likelihood Function}
\usage{
likbeta(x, shape1, shape2, ncp = 0, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape1}{non-negative parameters of the Beta distribution.}

\item{shape2}{non-negative parameters of the Beta distribution.}

\item{ncp}{non-centrality parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Beta Likelihood Function
}
