% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betalu.R
\name{likbetalu}
\alias{likbetalu}
\title{Betalu Likelihood Function}
\usage{
likbetalu(x, shape1, shape2, l = 0, u = 1, ncp = 0, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape1}{beta parameter shape1}

\item{shape2}{beta parameter shape2}

\item{l}{lower bound of support}

\item{u}{upper bound of support}

\item{ncp}{non-centrality parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Betalu Likelihood Function
}
