% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{likbinom}
\alias{likbinom}
\title{Binomial Likelihood Function}
\usage{
likbinom(x, size, prob, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{size}{number of trials (zero or more).}

\item{prob}{probability of success on each trial.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Binomial Likelihood Function
}
