% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood functions.R
\name{likcauchy}
\alias{likcauchy}
\title{Cauchy Likelihood Function}
\usage{
likcauchy(data, location, scale, log = TRUE, constants = FALSE)
}
\arguments{
\item{constants}{}
}
\description{
Cauchy Likelihood Function
}

