% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi.R
\name{likchi}
\alias{likchi}
\title{Chi Likelihood Function}
\usage{
likchi(x, df, ncp = 0, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{df}{degrees of freedom (non-negative, but can be
non-integer).}

\item{ncp}{non-centrality parameter (non-negative).}

\item{log}{logical; if TRUE, probabilities p are given as 
log(p).}
}
\description{
Chi Likelihood Function
}
