% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi-Squared.R
\name{likchisq}
\alias{likchisq}
\title{Chi-Squared Likelihood Function}
\usage{
likchisq(x, df, ncp = 0, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{df}{degrees of freedom (non-negative, but can be non-integer).}

\item{ncp}{non-centrality parameter (non-negative).}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Chi-Squared Likelihood Function
}
