% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\name{likdirichlet}
\alias{likdirichlet}
\title{Dirichlet Likelihood Function}
\usage{
likdirichlet(x, alpha, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{dirichlet alpha parameter (numeric vector with
positive entries)}

\item{log}{logical; if TRUE, probabilities p are given as 
log(p).}
}
\description{
Dirichlet Likelihood Function
}
