% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodExplore-package.R
\docType{package}
\name{likelihoodExplore}
\alias{likelihoodExplore}
\alias{likelihoodExplore-package}
\title{Likelihood Exploration}
\description{
A package of likelihood functions and a function that creates
likelihood functions from density functions. The functions are meant to aid
in education of likelihood based methods. Included shiny app leverages
likelihood functions to explore likelihood based methods.
}
