% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{likexp}
\alias{likexp}
\title{Exponential Likelihood Function}
\usage{
likexp(x, rate = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{vector of rates.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Exponential Likelihood Function
}
