% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{likf}
\alias{likf}
\title{F Likelihood Function}
\usage{
likf(x, df1, df2, ncp, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{df1}{degrees of freedom.  \code{Inf} is allowed.}

\item{df2}{degrees of freedom.  \code{Inf} is allowed.}

\item{ncp}{non-centrality parameter. If omitted the central F is assumed.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
F Likelihood Function
}
