% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{likgamma}
\alias{likgamma}
\title{Gamma Likelihood Function}
\usage{
likgamma(x, shape, rate = 1, scale = 1/rate, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{rate}{an alternative way to specify the scale.}

\item{scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{log}{logical; if \code{TRUE}, probabilities/densities \eqn{p}
    are returned as \eqn{log(p)}.}
}
\description{
Gamma Likelihood Function
}
