% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{likgeom}
\alias{likgeom}
\title{Geometric Likelihood Function}
\usage{
likgeom(x, prob, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles representing the number of failures in
    a sequence of Bernoulli trials before success occurs.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Geometric Likelihood Function
}
