% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypergeometric.R
\name{likhyper}
\alias{likhyper}
\title{Hypergeometric Likelihood Function}
\usage{
likhyper(x, m, n, k, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles representing the number of white balls
    drawn without replacement from an urn which contains both black and
    white balls.}

\item{m}{the number of white balls in the urn.}

\item{n}{the number of black balls in the urn.}

\item{k}{the number of balls drawn from the urn.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Hypergeometric Likelihood Function
}
