% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Normal.R
\name{liklnorm}
\alias{liklnorm}
\title{Log Normal Likelihood Function}
\usage{
liklnorm(x, meanlog = 0, sdlog = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{meanlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{sdlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Log Normal Likelihood Function
}
