% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{liklogis}
\alias{liklogis}
\title{Logistic Likelihood Function}
\usage{
liklogis(x, location = 0, scale = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Logistic Likelihood Function
}
