% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{liknorm}
\alias{liknorm}
\title{Normal Likelihood Function}
\usage{
liknorm(x, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Normal Likelihood Function
}
