% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{likpois}
\alias{likpois}
\title{Poisson Likelihood Function}
\usage{
likpois(x, lambda, log = TRUE)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{lambda}{vector of (non-negative) means.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Poisson Likelihood Function
}
