% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Student_t.R
\name{likt}
\alias{likt}
\title{Student t Likelihood Function}
\usage{
likt(x, df, ncp, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer).  \code{df
      = Inf} is allowed.}

\item{ncp}{non-centrality parameter \eqn{\delta}{delta};
    currently except for \code{rt()}, only for \code{abs(ncp) <= 37.62}.
    If omitted, use the central t distribution.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Student t Likelihood Function
}
