% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uniform.R
\name{likunif}
\alias{likunif}
\title{Uniform Likelihood Function}
\usage{
likunif(x, min = 0, max = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{min}{lower and upper limits of the distribution.  Must be finite.}

\item{max}{lower and upper limits of the distribution.  Must be finite.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Uniform Likelihood Function
}
