% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull.R
\name{likweibull}
\alias{likweibull}
\title{Weibull Likelihood Function}
\usage{
likweibull(x, shape, scale = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape}{shape and scale parameters, the latter defaulting to 1.}

\item{scale}{shape and scale parameters, the latter defaulting to 1.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Weibull Likelihood Function
}
