% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilcox.R
\name{likwilcox}
\alias{likwilcox}
\title{Wilcox Likelihood Function}
\usage{
likwilcox(x, m, n, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{m}{numbers of observations in the first and second sample,
    respectively.  Can be vectors of positive integers.}

\item{n}{numbers of observations in the first and second sample,
    respectively.  Can be vectors of positive integers.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Wilcox Likelihood Function
}
