% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{rbinomlik}
\alias{rbinomlik}
\title{Binomial Likelihood Function}
\usage{
rbinomlik(data, size, prob, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{size}{number of bernoulli trials}

\item{prob}{probability of success on each trial}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Binomial Likelihood Function
}
