% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{rcauchylik}
\alias{rcauchylik}
\title{Cauchy Likelihood Function}
\usage{
rcauchylik(data, location, scale, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{location}{location parameter}

\item{scale}{scale parameter}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Cauchy Likelihood Function
}
