% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi-Squared.R
\name{rchisqlik}
\alias{rchisqlik}
\title{Chi-Squared Likelihood Function}
\usage{
rchisqlik(data, df, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{df}{degrees of freedom}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Chi-Squared Likelihood Function
}

