% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{rexplik}
\alias{rexplik}
\title{Exponential Likelihood Function}
\usage{
rexplik(data, rate, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{rate}{rate of the exponential where the mean is 1 / rate}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Exponential Likelihood Function
}

