% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{rnormlik}
\alias{rnormlik}
\title{Normal Likelihood Function}
\usage{
rnormlik(data, ..., log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{...}{other options}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Normal Likelihood Function
}

