% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood functions.R
\name{likbeta}
\alias{likbeta}
\title{Beta Likelihood Function}
\usage{
likbeta(data, a, b, log = TRUE)
}
\arguments{
\item{data}{vector of data points}

\item{a}{shape parameter \alpha}

\item{b}{shape parameter \beta}

\item{log}{}
}
\description{
Beta Likelihood Function
}
\examples{
likbeta(rbeta(4, 0.5, 0.5), 0.5, 0.5)
}

