% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood functions.R
\name{likbinom}
\alias{likbinom}
\title{Binomial Likelihood Function}
\usage{
likbinom(data, size, prob, log = TRUE, constants = FALSE)
}
\arguments{
\item{log}{}
}
\description{
Binomial Likelihood Function
}

