% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{likcauchy}
\alias{likcauchy}
\title{Cauchy Likelihood Function}
\usage{
likcauchy(x, location = 0, scale = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Cauchy Likelihood Function
}
