% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood functions.R
\name{likchisq}
\alias{likchisq}
\title{Chi-Squared Likelihood Function}
\usage{
likchisq(data, df, non_centrality, log = TRUE, constants = FALSE)
}
\arguments{
\item{constants}{}
}
\description{
Chi-Squared Likelihood Function
}

