% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood Function}
\usage{
likelihood(density, ...)
}
\arguments{
\item{density}{density function used}

\item{...}{other options}

\item{data}{vector of data points}

\item{log}{logical if log values are used}
}
\description{
Likelihood Function
}
