% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Log Likelihood Function Maker}
\usage{
likelihood(density, ...)
}
\arguments{
\item{density}{density function used}

\item{...}{other options}
}
\value{
A function that is the log likelihood function from density function
density.
}
\description{
Function that creates a log likelihood function given a density function
density.
}
\details{
The log likelihood is the log of a function of parameters given the data.
}
\examples{
likelihood(dnorm, x = rnorm(100))

}
