% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{likexp}
\alias{likexp}
\title{Exponential Log Likelihood Function}
\usage{
likexp(x, rate = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{vector of rates.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric scalar for the log likelihood of the exponential density
given the data where rate can be held constant or if vector were given
vector will be returned.
}
\description{
The log likelihood of a exponential density with data, x, rate parameter.
}
\details{
The log likelihood is the log of a function of parameters given the data.
}
\examples{
likexp(x = rexp(n = 2))
}
