% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_Gamma.R
\name{likinvgamma}
\alias{likinvgamma}
\title{Inverse Gamma Likelihood Function}
\usage{
likinvgamma(x, shape, rate = 1, scale = 1/rate, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape}{inverse gamma shape parameter}

\item{rate}{inverse gamma rate parameter}

\item{scale}{alternative to rate; scale = 1/rate}

\item{log}{logical; if TRUE, probabilities p are given as
log(p).}
}
\description{
Inverse Gamma Likelihood Function
}
