% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Normal.R
\name{liklnorm}
\alias{liklnorm}
\title{Log Normal Log Likelihood Function}
\usage{
liklnorm(x, meanlog = 0, sdlog = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{meanlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{sdlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric scalar for the log likelihood of the log normal density
given the data where meanlog and sdlog can be held constant or if vector were
given vector will be returned.
}
\description{
The log likelihood of a log normal density with data, x, meanlog and sdlog
parameters.
}
\details{
The log likelihood is the log of a function of parameters given the data.
}
\examples{
liklnorm(x = rlnorm(n = 2))
}
