% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{liklogis}
\alias{liklogis}
\title{Logistic Log Likelihood Function}
\usage{
liklogis(x, location = 0, scale = 1, log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric scalar for the log likelihood of the logistic density given
the data where location and scale can be held constant or if vector were given
vector will be returned.
}
\description{
The log likelihood of a logistic density with data, x, location and scale
parameters.
}
\details{
The log likelihood is the log of a function of parameters given the data.
}
\examples{
liklogis(x = rlogis(n = 2))
}
