% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood functions.R
\name{liknorm}
\alias{liknorm}
\title{Normal Likelihood Function}
\usage{
liknorm(data, mu, sigma, log = TRUE, constants = FALSE)
}
\arguments{
\item{data}{vector of data points}

\item{mu}{mean}

\item{sigma}{standard deviation}

\item{log}{logical; if TRUE, output is the log likelihood}
}
\description{
Normal Likelihood Function
}
\examples{
liknorm(rnorm(4), 0, 1)
}

